- create Claimer object - move business logic out of main - fix unintended concurrent mnemonic usage
