+++ deputy-claimer/go.mod
@@ -8,5 +8,4 @@ require (
 	github.com/kava-labs/go-sdk v0.1.6
 	github.com/kava-labs/tendermint v0.33.4-0.20200520221629-77480532c622
 	github.com/stretchr/testify v1.5.1
-	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e
 )
